Cleanup Project Layout (#159)
