--- busy_beaver/decorators/__init__.py
@@ -1,1 +1,0 @@
-from .authenication import authentication_required  # noqa
