+++ busy_beaver/toolbox/rq.py
@@ -1,4 +1,5 @@
 import logging
+from rq import get_current_job
 
 from busy_beaver.config import TASK_QUEUE_MAX_RETRIES
 from busy_beaver.exceptions import AsyncException
@@ -7,6 +8,19 @@ from busy_beaver.models import Task
 
 logger = logging.getLogger(__name__)
 MAX_FAILURES = TASK_QUEUE_MAX_RETRIES + 1
+
+
+def set_task_progress(progress):
+    job = get_current_job()
+    if job:
+        job.meta["progress"] = progress
+        job.save_meta()
+
+        if progress >= 100:
+            task = Task.query.filter_by(job_id=job.get_id()).first()
+            if task:
+                task.complete = True
+                db.session.commit()
 
 
 def retry_failed_job(job, *exc_info):
